requirements.txt
pymongo
python-dotenv
web3>=6.0.0
requests>=2.31.0
